fix golint warnning
