+++ pkg/yurthub/storage/factory/factory.go
@@ -5,6 +5,7 @@ import (
 	"github.com/alibaba/openyurt/pkg/yurthub/storage/disk"
 )
 
+// CreateStorage create a storage.Store for backend storage
 func CreateStorage() (storage.Store, error) {
 
 	return disk.NewDiskStorage()
