+++ pkg/yurthub/storage/store.go
@@ -4,8 +4,10 @@ import (
 	"errors"
 )
 
+// ErrStorageAccessConflict is an error for accessing key conflict
 var ErrStorageAccessConflict = errors.New("specified key is under accessing")
 
+// Store is an interface for caching data into backend storage
 type Store interface {
 	Create(key string, contents []byte) error
 	Delete(key string) error
